+++ code/packages/sdk/src/actions/frame-tool/constants.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 2025 INDUSTRIA DE DISEÑO TEXTIL S.A. (INDITEX S.A.)
+//
+// SPDX-License-Identifier: Apache-2.0
+
 export const FRAME_TOOL_ACTION_NAME = 'frameTool';
 
 export const FRAME_TOOL_STATE = {