+++ code/packages/sdk/src/actions/rectangle-tool/constants.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 2025 INDUSTRIA DE DISEÑO TEXTIL S.A. (INDITEX S.A.)
+//
+// SPDX-License-Identifier: Apache-2.0
+
 export const RECTANGLE_TOOL_ACTION_NAME = 'rectangleTool';
 
 export const RECTANGLE_TOOL_STATE = {