+++ code/packages/sdk/src/plugins/stage-resize/stage-resize.ts
@@ -1,4 +1,8 @@
-import { WeavePlugin } from '@/plugins/plugin';
+// SPDX-FileCopyrightText: 2025 2025 INDUSTRIA DE DISEÑO TEXTIL S.A. (INDITEX S.A.)
+//
+// SPDX-License-Identifier: Apache-2.0
+
+import { WeavePlugin } from "@/plugins/plugin";
 
 export class WeaveStageResizePlugin extends WeavePlugin {
   getLayerName = undefined;