+++ code/packages/sdk/src/weave.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 2025 INDUSTRIA DE DISEÑO TEXTIL S.A. (INDITEX S.A.)
+//
+// SPDX-License-Identifier: Apache-2.0
+
 import Emittery from 'emittery';
 import Konva from 'konva';
 import { Vector2d } from 'konva/lib/types';