+++ code/packages/store-websockets/src/constants.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 2025 INDUSTRIA DE DISEÑO TEXTIL S.A. (INDITEX S.A.)
+//
+// SPDX-License-Identifier: Apache-2.0
+
 export const WEAVE_STORE_WEBSOCKETS = "store-websockets";
 
 export const WEAVE_STORE_WEBSOCKETS_CONNECTION_STATUS = {