+++ code/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.20.0] - 2025-04-07
+## [0.20.1] - 2025-04-07
 
 - [#74](https://github.com/InditexTech/weavejs/issues/74) Improve grid plugin performance